[observability] Split out httpAPI and httpViewer instrumentation
Split out the Observability interface so that API route and viewer route are dealt with separately.

Moved Null Observer to an internal package, treating nil as a Null Observer.
Prometheus observer can be configured to indicate which API vars should be filled in. Default variables
are :map_name, :layer_name and :z.